# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PURCHASES ---
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List automation workflows
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workflows/{id}/trigger:
    post:
      operationId: triggerWorkflow
      tags: [workflows]
      summary: Manually trigger a workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRun"

  # --- NOTIFICATIONS ---
...
components:
  schemas:
...
    Vendor:
      type: object
      required: [id, name, email, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        status:
          type: string
        category:
          type: ["string", "null"]
    VendorInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
        category:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
        category:
          type: string

    # --- PURCHASES ---
...
    Workflow:
      type: object
      required: [id, name, trigger, status, runs, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        trigger:
          type: string
        status:
          type: string
        runs:
          type: integer
        last_run:
          type: ["string", "null"]
        created_at:
          type: string
    WorkflowInput:
      type: object
      required: [name, trigger]
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        status:
          type: string
    WorkflowUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: string
        status:
          type: string
    WorkflowRun:
      type: object
      required: [workflow_id, status, timestamp]
      properties:
        workflow_id:
          type: integer
        status:
          type: string
        timestamp:
          type: string
        message:
          type: string

    # --- NOTIFICATIONS ---
...